+++ src/app/management/components/header/management-header.component.ts
@@ -10,7 +10,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { filter, pluck, tap } from 'rxjs/operators';
+import { filter, tap } from 'rxjs/operators';
 import { AppService } from 'src/app/app.service';
 import { ManagementService } from 'src/app/management/management.service';
 import { EditProfileFormDialogComponent } from 'src/app/shared/dialogs/edit-profile-form/edit-profile-form-dialog.component';
@@ -39,7 +39,7 @@ export class ManagementHeaderComponent {
     protected snackBarService: MatSnackBar
   ) {
     this.moduleName$ = this.service.currentPageName$.pipe();
-    this.userName$ = this.appService.getUserProfile().pipe(pluck('name'));
+    this.userName$ = this.appService.userName$.pipe();
   }
 
   switchSidenavOpenState(): void {